+++ PhpSerializerNET.Test/Deserialize/ObjectDeserialization.cs
@@ -6,6 +6,7 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
 **/
 
 using Xunit;
+using System.Collections.Generic;
 using PhpSerializerNET.Test.DataTypes;
 
 namespace PhpSerializerNET.Test.Deserialize;
@@ -22,5 +23,25 @@ public class ObjectDeserializationTest {
 		Assert.Equal("Bar", result.Bar);
 		Assert.Equal("A", result.Baz);
 		Assert.Equal("B", result.Dummy);
+	}
+
+	[Fact]
+	public void ListOfObjects() {
+		// Regression test for https://github.com/StringEpsilon/PhpSerializerNET/issues/40
+		var result = PhpSerialization.Deserialize<List<MixedKeysPhpClass>>(
+			"""a:3:{i:0;a:4:{s:3:"Foo";s:3:"Foo";s:3:"Bar";s:3:"Bar";s:1:"a";s:1:"A";s:1:"b";s:1:"B";}i:1;a:4:{s:3:"Foo";s:3:"Foo";s:3:"Bar";s:3:"Bar";s:1:"a";s:1:"A";s:1:"b";s:1:"B";}i:2;a:4:{s:3:"Foo";s:3:"Foo";s:3:"Bar";s:3:"Bar";s:1:"a";s:1:"A";s:1:"b";s:1:"B";}}"""
+		);
+
+		Assert.Equal(3, result.Count);
+	}
+	[Fact]
+	public void ImplicitListOfObjects() {
+		// Regression test for https://github.com/StringEpsilon/PhpSerializerNET/issues/40
+		var result = PhpSerialization.Deserialize(
+			"""a:3:{i:0;a:4:{s:3:"Foo";s:3:"Foo";s:3:"Bar";s:3:"Bar";s:1:"a";s:1:"A";s:1:"b";s:1:"B";}i:1;a:4:{s:3:"Foo";s:3:"Foo";s:3:"Bar";s:3:"Bar";s:1:"a";s:1:"A";s:1:"b";s:1:"B";}i:2;a:4:{s:3:"Foo";s:3:"Foo";s:3:"Bar";s:3:"Bar";s:1:"a";s:1:"A";s:1:"b";s:1:"B";}}""",
+			new PhpDeserializationOptions { UseLists = ListOptions.Default }
+		) as List<object>;
+
+		Assert.Equal(3, result.Count);
 	}
 }